Move responses to application
