+++ applications/api/src/Infrastructure/Serializer/ApiSerializerDriver.php
@@ -2,8 +2,8 @@
 
 namespace CodelyTv\Api\Infrastructure\Serializer;
 
-use CodelyTv\Context\Video\Module\User\Application\UserResponse;
-use CodelyTv\Context\Video\Module\Video\Domain\VideoResponse;
+use CodelyTv\Context\Video\Module\User\Application\Find\UserResponse;
+use CodelyTv\Context\Video\Module\Video\Application\Find\VideoResponse;
 use CodelyTv\Infrastructure\Jms\CodelyTvSerializerDriver;
 
 final class ApiSerializerDriver extends CodelyTvSerializerDriver
