+++ src/Context/Video/Module/Video/Application/Find/VideoResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Context\Video\Module\Video\Domain;
+namespace CodelyTv\Context\Video\Module\Video\Application\Find;
 
 use CodelyTv\Shared\Domain\Bus\Query\Response;
