+++ src/Context/Video/Module/Video/Application/Find/VideoResponseConverter.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Context\Video\Module\Video\Domain;
+namespace CodelyTv\Context\Video\Module\Video\Application\Find;
+
+use CodelyTv\Context\Video\Module\Video\Domain\Video;
 
 final class VideoResponseConverter
 {
