+++ tests/Context/Video/Module/User/Domain/UserResponseStub.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Test\Context\Video\Module\User\Domain;
 
-use CodelyTv\Context\Video\Module\User\Application\UserResponse;
+use CodelyTv\Context\Video\Module\User\Application\Find\UserResponse;
 use CodelyTv\Context\Video\Module\User\Domain\TotalVideosCreated;
 use CodelyTv\Context\Video\Module\User\Domain\UserId;
 use CodelyTv\Context\Video\Module\User\Domain\UserName;
