+++ tests/Context/Video/Module/Video/Application/Find/VideoResponseStub.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace CodelyTv\Test\Context\Video\Module\Video\Application\Find;
 
+use CodelyTv\Context\Video\Module\Video\Application\Find\VideoResponse;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
-use CodelyTv\Context\Video\Module\Video\Domain\VideoResponse;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoTitle;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoType;
 use CodelyTv\Context\Video\Module\Video\Domain\VideoUrl;
